refactor(ImageFeed): Финальный рефакторинг
META: sprint_13, ImageFeed
