--- ImageFeed/ImageFeed/Protocols/ImagesListCellDelegate.swift
@@ -1,3 +1,0 @@
-protocol ImagesListCellDelegate: AnyObject {
-    func didTapLike(_ cell: ImagesListCell)
-}
